- Created infrastructure for BMIntefaceGameAction - Moved create_game from BMInterface to BMInterfaceGameAction
