+++ test/src/engine/BMInterfacePlayerTest.php
@@ -133,9 +133,12 @@ class BMInterfacePlayerTest extends BMInterfaceTestAbstract {
      * @covers BMInterfacePlayer::update_last_action_time
      */
     public function test_update_last_action_time() {
-        $retval = $this->object->create_game(array(self::$userId1WithoutAutopass,
-                                                   self::$userId2WithoutAutopass),
-                                                   array('Avis', 'Hammer'), 4);
+        $retval = $this->object->gameAction()->create_game(
+            array(self::$userId1WithoutAutopass,
+                  self::$userId2WithoutAutopass),
+            array('Avis', 'Hammer'),
+            4
+        );
         $gameId = $retval['gameId'];
 
         $game = self::load_game($gameId);
